Guy Raviv 26.1.16 - Remove invalid XML 1.0 characters from Exchange XML responses
We will get this exception if the xml will contain non-valid characters.
Clients can send emails with non-valid characters but EWS library will pass them as-is - without cleaning/replacing
them. To solve this, we will make another request and we will wrap the response InputStream with our ReplacingInputStream
which will remove (maybe at the future we will do replacements) the non-valid characters.